Make it more clear that users *need* an execution client.
With the merge closing in on it, anyone setting up a client at this time should not be using a third party execution layer.  I removed the optionality, removed references to third party data providers, and removed the optionality of execution clients.

This has PR #149 baked in, feel free to close that one if this is merged first, or I can rebase if that one is merged first.